Added LoginFunction in controller

+++ Snake/SDK/User.java
@@ -7,7 +7,7 @@ import java.sql.Date;
  */
 public class User {
 
-    private int id;
+    private long id;
     private String first_name;
     private String last_name;
     private String email;
@@ -19,12 +19,12 @@ public class User {
 
     public User()
     {}
-        public int getId()
+        public long getId()
         {
             return id;
         }
 
-        public void setId(int _id)
+        public void setId(long _id)
     {
         id = _id;
     }
